move integration tests to the integration test repo (DAT-14969) (#4341)
Reenables most of the integration tests disabled in #4339, and moves them into the integration-tests module.
