+++ liquibase-standard/src/test/groovy/liquibase/LiquibaseTest.groovy
@@ -263,221 +263,6 @@ class LiquibaseTest extends Specification {
 //        });
 //    }
 
-
-    def syncChangeLogForUnmanagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-
-        Liquibase liquibase = createUnmanagedDatabase(h2Connection);
-        assertFalse(hasDatabaseChangeLogTable(liquibase));
-
-        liquibase.changeLogSync("");
-
-        then:
-        assert hasDatabaseChangeLogTable(liquibase);
-        assertTags(liquibase, "1.0", "1.1", "2.0");
-    }
-
-    def syncChangeLogToTagForUnmanagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-
-        Liquibase liquibase = createUnmanagedDatabase(h2Connection);
-
-        then:
-        assert !hasDatabaseChangeLogTable(liquibase)
-
-        when:
-        liquibase.changeLogSync("1.1", "");
-
-        then:
-        assert hasDatabaseChangeLogTable(liquibase);
-        assertTags(liquibase, "1.0", "1.1");
-    }
-
-    def syncChangeLogForManagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-
-        Liquibase liquibase = createDatabaseAtTag(h2Connection, "1.0");
-
-        then:
-        assert hasDatabaseChangeLogTable(liquibase)
-
-        when:
-        liquibase.changeLogSync("");
-
-        then:
-        assertTags(liquibase, "1.0", "1.1", "2.0");
-    }
-
-    def syncChangeLogToTagForManagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-
-        Liquibase liquibase = createDatabaseAtTag(h2Connection, "1.0");
-        then:
-        assert hasDatabaseChangeLogTable(liquibase);
-
-        when:
-        liquibase.changeLogSync("1.1", "");
-
-        then:
-        assertTags(liquibase, "1.0", "1.1");
-    }
-
-    def syncChangeLogSqlForUnmanagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-        StringWriter writer = new StringWriter();
-
-        Liquibase liquibase = createUnmanagedDatabase(h2Connection);
-
-        then:
-        assert !hasDatabaseChangeLogTable(liquibase);
-
-        when:
-        liquibase.changeLogSync("", writer);
-
-        then:
-        assert !hasDatabaseChangeLogTable(liquibase);
-        assertSqlOutputAppliesTags(writer.toString(), "1.0", "1.1", "2.0");
-    }
-
-    def syncChangeLogToTagSqlForUnmanagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-        StringWriter writer = new StringWriter();
-
-        Liquibase liquibase = createUnmanagedDatabase(h2Connection);
-
-        then:
-        assert !hasDatabaseChangeLogTable(liquibase);
-
-        when:
-        liquibase.changeLogSync("1.1", "", writer);
-
-        then:
-        !hasDatabaseChangeLogTable(liquibase);
-        assertSqlOutputAppliesTags(writer.toString(), "1.0", "1.1");
-    }
-
-    def syncChangeLogSqlForManagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-        StringWriter writer = new StringWriter();
-
-        Liquibase liquibase = createDatabaseAtTag(h2Connection, "1.0");
-
-        then:
-        assert hasDatabaseChangeLogTable(liquibase);
-
-        when:
-        liquibase.changeLogSync("", writer);
-
-        then:
-        assertSqlOutputAppliesTags(writer.toString(), "1.1", "2.0");
-    }
-
-    def syncChangeLogToTagSqlForManagedDatabase() throws Exception {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection();
-        StringWriter writer = new StringWriter();
-
-        Liquibase liquibase = createDatabaseAtTag(h2Connection, "1.0");
-
-        then:
-        assertTrue(hasDatabaseChangeLogTable(liquibase));
-
-        when:
-        liquibase.changeLogSync("1.1", "", writer);
-
-        then:
-        assertSqlOutputAppliesTags(writer.toString(), "1.1");
-    }
-
-    @Ignore
-    def "validate checksums from ran changesets have all been reset"() {
-        when:
-        h2Connection = getInMemoryH2DatabaseConnection()
-        Liquibase liquibase = new Liquibase("liquibase/test-changelog-fast-check.xml", new ClassLoaderResourceAccessor(),
-                h2Connection)
-        liquibase.update()
-        liquibase.clearCheckSums()
-
-        then:
-        List<RanChangeSet> ranChangeSets = ChangeLogHistoryServiceFactory.getInstance().getChangeLogService(liquibase.getDatabase()).getRanChangeSets()
-        assert ranChangeSets.get(0).getLastCheckSum() == null
-    }
-
-
-    private JdbcConnection getInMemoryH2DatabaseConnection() throws SQLException {
-        String urlFormat = "jdbc:h2:mem:%s";
-        return new JdbcConnection(DriverManager.getConnection(format(urlFormat, UUID.randomUUID().toString())));
-    }
-
-    private Liquibase createUnmanagedDatabase(JdbcConnection connection) throws SQLException, LiquibaseException {
-        String createTableSql = "CREATE TABLE PUBLIC.TABLE_A (ID INTEGER);";
-
-        PreparedStatement stmt = connection.getUnderlyingConnection().prepareStatement(createTableSql)
-        try {
-            stmt.execute();
-        } finally {
-            stmt.close()
-        }
-
-        return new Liquibase("liquibase/tagged-changelog.xml", new ClassLoaderResourceAccessor(), connection);
-    }
-
-    private Liquibase createDatabaseAtTag(JdbcConnection connection, String tag) throws LiquibaseException {
-        Liquibase liquibase = new Liquibase("liquibase/tagged-changelog.xml", new ClassLoaderResourceAccessor(),
-                connection);
-        liquibase.update(tag, "");
-        return liquibase;
-    }
-
-    private boolean hasDatabaseChangeLogTable(Liquibase liquibase) throws DatabaseException {
-        return SnapshotGeneratorFactory.getInstance().hasDatabaseChangeLogTable(liquibase.database);
-    }
-
-    private void assertTags(Liquibase liquibase, String... expectedTags) throws DatabaseException {
-        def actualTags = []
-        for (def ranChangeset : liquibase.database.getRanChangeSetList()) {
-            if (ranChangeset.getTag() != null) {
-                actualTags.add(ranChangeset.getTag())
-            }
-        }
-
-        assertEquals(Arrays.asList(expectedTags), actualTags);
-    }
-
-    private void assertSqlOutputAppliesTags(String output, String... expectedTags) throws IOException {
-        String insertTagH2SqlTemplate =
-                "INSERT INTO PUBLIC\\.DATABASECHANGELOG \\(.*, DESCRIPTION,.*, TAG\\) VALUES \\(.*, 'tagDatabase',.*, '%s'\\);";
-
-        List<Pattern> patterns = []
-
-        for (def tag : expectedTags) {
-            patterns.add(Pattern.compile(String.format(insertTagH2SqlTemplate, tag)))
-        }
-
-        BufferedReader reader = new BufferedReader(new StringReader(output))
-        try {
-            String line;
-            int index = 0;
-
-            while ((line = reader.readLine()) != null && index < patterns.size()) {
-                Matcher matcher = patterns.get(index).matcher(line);
-                if (matcher.matches()) {
-                    index++;
-                }
-            }
-            assertTrue(index > 0 && index == patterns.size());
-        } finally {
-            reader.close()
-        }
-    }
-
     static class TestConsoleUIService extends ConsoleUIService {
         private List<String> messages = new ArrayList<>()
